+++ src/algorithms/constructors/mod.rs
@@ -14,6 +14,7 @@ pub mod hypercube;
 pub mod kary_tree;
 pub mod kautz;
 pub mod linegraph;
+pub mod prufer;
 pub mod regular_tree;
 pub mod ring;
 pub mod square_lattice;