+++ src/lib.rs
@@ -110,6 +110,7 @@ pub use crate::algorithms::constructors::hypercube::{MAX_HYPERCUBE_DIMENSION, hy
 pub use crate::algorithms::constructors::kary_tree::{TreeMode, kary_tree};
 pub use crate::algorithms::constructors::kautz::kautz;
 pub use crate::algorithms::constructors::linegraph::linegraph;
+pub use crate::algorithms::constructors::prufer::from_prufer;
 pub use crate::algorithms::constructors::regular_tree::regular_tree;
 pub use crate::algorithms::constructors::ring::{cycle_graph, path_graph, ring_graph};
 pub use crate::algorithms::constructors::square_lattice::square_lattice;